feat: Allow subscribers/publishers to reuse a single connection
A subscriber and publisher may now be created with a shared connection, i.e. multiple subscribers and publishers may use the same connection. A new option is added to NewSubscriber and NewPublisher: WithConnection. The existing behavior is maintained, i.e. when a connection is not supplied then a new connection is created for the subscriber/publisher (as before), and when the subscriber/publisher is closed then the connection is also closed. If a connection is provided to the subscriber/publisher, then when Close is called on the subscriber/publisher, only the resources of the specific subscriber/publisher are cleaned up. In the case of a subscriber, the subscription message listeners are stopped and all message channels are closed, but the connection is not closed. The client that created the connection is responsible for closing the connection.

closes #182

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>