feat: Implement a publisher channel pool
A new Publisher configuration parameter was added: ChannelPoolSize. If this parameter is set to 0 (default) then there is no change to the existing behavior, i.e. a channel is opened/closed for every Publish operation. If this parameter is set to a value greater than 0 then a channel pool is maintained, where channels are reused. The number of channels for a publisher will not exceed the value of this parameter, i.e. if the Publish function is invoked and there are no available channels in the pool then the caller will wait.

The channel pool improves performance in high volume systems by not incurring the overhead of opening and closing channels, and the pool also acts as a throttle to prevent the AMQP server from overloading.

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>