fix: Deadlock in subscriber consuming loop
When an AMQP server is restarted then the subscriber channels are closed and 'NotifyClose' handlers are notified. Currently the Go channel registered with the AMQP channel has no buffer and this intermittently causes a deadlock. For example, if an error occurs while processing a message in the ConsumingLoop then the function exits, so when the AMQP channel attempts to send an error to the Go channel when it is being shut down, there are no consumers and a deadlock occurs. The deadlock is prevented by adding a buffer to the Go channel.

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>